+++ tests/test_autoimage.py
@@ -2,16 +2,13 @@ import unittest
 import pytraj as pt
 from pytraj.base import *
 from pytraj import adict
-from pytraj import io as mdio
-import pytraj.io as io
-from pytraj.utils.check_and_assert import assert_almost_equal
-from pytraj.decorators import test_if_having, no_test
-from pytraj.compat import izip as zip
+from pytraj.testing import aa_eq, test_if_having
+from pytraj.compat import zip
 
 
-class Test(unittest.TestCase):
+class TestRegular(unittest.TestCase):
     def test_1(self):
-        traj = mdio.iterload(
+        traj = pt.iterload(
             "./data/tz2.truncoct.nc", "./data/tz2.truncoct.parm7")
         f0 = traj[0]
         f0cp = f0.copy()
@@ -21,14 +18,14 @@ class Test(unittest.TestCase):
         #print(f0.same_coords_as(f0cp))
         assert f0.same_coords_as(f0cp) == False
 
-        fsaved = mdio.iterload("./data/tz2.truncoct.autoiamge.save.r",
+        fsaved = pt.iterload("./data/tz2.truncoct.autoiamge.save.r",
                                "./data/tz2.truncoct.parm7")[0]
-        assert_almost_equal(fsaved.coords, f0.coords)
+        aa_eq(fsaved.coords, f0.coords)
 
     def test_2(self):
         from pytraj.common_actions import do_autoimage
         # test do_autoimage
-        traj = mdio.iterload(
+        traj = pt.iterload(
             "./data/tz2.truncoct.nc", "./data/tz2.truncoct.parm7")
         f0 = traj[0]
         f0cp = f0.copy()
@@ -38,47 +35,14 @@ class Test(unittest.TestCase):
         #print(f0.same_coords_as(f0cp))
         assert f0.same_coords_as(f0cp) == False
 
-        fsaved = mdio.iterload("./data/tz2.truncoct.autoiamge.save.r",
+        fsaved = pt.iterload("./data/tz2.truncoct.autoiamge.save.r",
                                "./data/tz2.truncoct.parm7")[0]
-        assert_almost_equal(fsaved.coords, f0.coords)
-
-    @no_test
-    @test_if_having("mdtraj")
-    def test_3(self):
-        import mdtraj as md
-        import numpy as np
-        # load from mdtraj
-        # test do_autoimage
-        traj = mdio.iterload(
-            "./data/tz2.truncoct.nc", "./data/tz2.truncoct.parm7")
-        fa = traj[:]
-        for f0, f1 in zip(traj, fa):
-            assert f0.box.type == traj.top.box.type == f1.box.type == 'truncoct'
-        m_traj = md.load_netcdf(
-            "./data/tz2.truncoct.nc",
-            top="./data/tz2.truncoct.parm7")
-        # create mutable Trajectory from TrajectoryIterator
-        fake_fa = io.load_mdtraj(m_traj)
-        for frame in fake_fa:
-            assert frame.box.type == 'truncoct'
-        assert_almost_equal(fa.xyz.flatten(), traj.xyz.flatten())
-        assert_almost_equal(fa.xyz.flatten(), fake_fa.xyz.flatten())
-        # do autoiamge
-        fa.autoimage()
-        #print(fa[0, 0])
-        # try loading from `mdtraj` object
-        fake_fa.autoimage()
-        #print(fake_fa[0, 0])
-        assert_almost_equal(fa.xyz.flatten(), fake_fa.xyz.flatten(), decimal=1)
-
-        for frame in fake_fa:
-            assert frame.has_box() == True
-        assert fake_fa.top.has_box() == True
+        aa_eq(fsaved.coords, f0.coords)
 
     def test_4(self):
         # combined with get_coordinates
         from pytraj.misc import rmsd_1darray
-        traj0 = mdio.iterload(
+        traj0 = pt.iterload(
             "./data/tz2.truncoct.nc", "./data/tz2.truncoct.parm7")
 
         # test autoimage
@@ -117,6 +81,38 @@ class Test(unittest.TestCase):
         xyz1 = traj1['@CA,C,N'].xyz
         # PASSED
         assert rmsd_1darray(xyz1.flatten(), xyz2.flatten()) < 1E-10
+
+
+class TestWithRmsfit(unittest.TestCase):
+    def test_0(self):
+        # TrajectoryIterrator
+        # status: failed
+        from pytraj.compat import zip
+        traj = pt.iterload("./data/tz2.ortho.nc", "./data/tz2.ortho.parm7")
+        pt.write_traj("./output/tz2.autoimage_with_rmsfit.nc",
+                      traj(autoimage=True,
+                           rmsfit=(0, '@CA,C,N')),
+                      overwrite=True)
+
+        saved_traj = pt.load('data/tz2.autoimage_with_rmsfit.nc', traj.top)
+        p_traj = pt.load('./output/tz2.autoimage_with_rmsfit.nc', traj.top)
+
+        aa_eq(saved_traj.xyz, p_traj.xyz)
+        for f1, f2 in zip(p_traj, saved_traj):
+            pass
+
+    def test_1(self):
+        # Trajectory (mutable)
+        # status: OK
+        from pytraj.compat import zip
+        # note: use `load` instead of `iterload`
+        traj = pt.load("./data/tz2.ortho.nc", "./data/tz2.ortho.parm7")
+        traj.autoimage()
+        traj.rmsfit(mask='@CA,C,N')
+        saved_traj = pt.load('data/tz2.autoimage_with_rmsfit.nc', traj.top)
+
+        # PASSED
+        aa_eq(saved_traj.xyz, traj.xyz)
 
 
 if __name__ == "__main__":
